Moves the attention calculation to attention.py for better code structure.
